refactor(chat): collapse getChatMcpTools tool wrappers into chat-tool-builder
