+++ platform/backend/src/clients/chat-mcp-client.test.ts
@@ -15,7 +15,11 @@ import ToolModel from "@/models/tool";
 import { resolveSessionExternalIdpToken } from "@/services/identity-providers/session-token";
 import { describe, expect, test } from "@/test";
 import * as chatClient from "./chat-mcp-client";
-import { mcpToolToModelOutput } from "./chat-mcp-client";
+import {
+  buildArchestraToolOutput,
+  mcpToolToModelOutput,
+  __test as toolBuilderTest,
+} from "./chat-tool-builder";
 import mcpClient from "./mcp-client";
 
 const mockConnect = vi.fn().mockRejectedValue(new Error("Connection closed"));
@@ -92,7 +96,7 @@ describe("isBrowserMcpTool", () => {
 });
 
 describe("normalizeJsonSchema", () => {
-  const { normalizeJsonSchema } = chatClient.__test;
+  const { normalizeJsonSchema } = toolBuilderTest;
 
   test("returns fallback schema for missing/invalid input", () => {
     expect(normalizeJsonSchema(null)).toEqual({
@@ -520,7 +524,7 @@ describe("executeMcpTool error handling", () => {
       }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
     expect(result.content).toBe("Auth required: install the server");
     expect(result._meta).toMatchObject({
       archestraError: expect.objectContaining({
@@ -549,7 +553,7 @@ describe("executeMcpTool error handling", () => {
       }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
     expect(result.content).toBe("Error line 1\nError line 2");
   });
 
@@ -560,7 +564,7 @@ describe("executeMcpTool error handling", () => {
       }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
     expect(result.content).toBe(
       JSON.stringify({ type: "image", data: "base64..." }),
     );
@@ -571,7 +575,7 @@ describe("executeMcpTool error handling", () => {
       mockResult({ content: null, error: "Something failed" }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
     expect(result.content).toBe("Something failed");
   });
 
@@ -580,7 +584,7 @@ describe("executeMcpTool error handling", () => {
       mockResult({ content: null }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
     expect(result.content).toBe("Tool execution failed");
   });
 
@@ -618,7 +622,7 @@ describe("executeMcpTool error handling", () => {
       }),
     );
 
-    const result = await chatClient.__test.executeMcpTool(baseCtx);
+    const result = await toolBuilderTest.executeMcpTool(baseCtx);
 
     expect(result._meta).toMatchObject({
       archestraError: expect.objectContaining({
@@ -660,7 +664,7 @@ describe("executeMcpTool error handling", () => {
       action: "mark_as_untrusted",
     });
 
-    const result = await chatClient.__test.executeMcpTool({
+    const result = await toolBuilderTest.executeMcpTool({
       ...baseCtx,
       agentId: agent.id,
       globalToolPolicy: "restrictive",
@@ -1589,7 +1593,7 @@ describe("buildArchestraToolOutput", () => {
     makeAgent,
   }) => {
     const agent = await makeAgent();
-    const result = await chatClient.buildArchestraToolOutput({
+    const result = await buildArchestraToolOutput({
       response: archestraResponse,
       toolName: "archestra__whoami",
       toolArguments: {},
@@ -1612,7 +1616,7 @@ describe("buildArchestraToolOutput", () => {
       // biome-ignore lint/suspicious/noExplicitAny: test mock data
       .mockResolvedValueOnce(mockToolDef as any);
 
-    const result = await chatClient.buildArchestraToolOutput({
+    const result = await buildArchestraToolOutput({
       response: archestraResponse,
       toolName: "archestra__run_tool",
       toolArguments: {
@@ -1645,7 +1649,7 @@ describe("buildArchestraToolOutput", () => {
       // biome-ignore lint/suspicious/noExplicitAny: test mock data
       .mockResolvedValueOnce({ name: "context7__search", meta: null } as any);
 
-    const result = await chatClient.buildArchestraToolOutput({
+    const result = await buildArchestraToolOutput({
       response: archestraResponse,
       toolName: "archestra__run_tool",
       toolArguments: { tool_name: "context7__search", tool_args: {} },
@@ -1660,7 +1664,7 @@ describe("buildArchestraToolOutput", () => {
 
 describe("throwIfApprovalRequired", () => {
   const { resolveApprovalPolicyTarget, throwIfApprovalRequired } =
-    chatClient.__test;
+    toolBuilderTest;
 
   test("does not throw when globalToolPolicy is permissive", async () => {
     await expect(
